DNA model creation partial refactor
DNA model creation simplified to not have to use str model names everytime.
Moved K80-related functions/structures to separate module, added pub use statement to
dna_models module
